Expose daml-helper as a library
